# backend-go-payment/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Payment processing service for e-ticketing platform",
        "title": "Payment Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/api/v1/payments": {
            "post": {
                "description": "Process a payment transaction for an order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Make Payment",
                "parameters": [
                    {
                        "description": "Payment Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.PaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/transactions": {
            "get": {
                "description": "Retrieve all payment transactions with pagination support",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get all transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/order/{order_id}": {
            "get": {
                "description": "Retrieve a transaction by its associated order_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transaction by order ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/txn/{transaction_id}": {
            "get": {
                "description": "Retrieve a transaction by its unique transaction_id (e.g., TXN-CC-17000000)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transaction by transaction ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID (e.g., TXN-CC-17000000)",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/{id}": {
            "get": {
                "description": "Retrieve a specific transaction by its primary key ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transaction by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.PaymentRequest": {
            "type": "object",
            "required": [
                "amount",
                "order_id",
                "payment_method"
            ],
            "properties": {
                "amount": {
                    "description": "Payment amount\nrequired: true\nminimum: 0.01\nexample: 250.50",
                    "type": "number"
                },
                "card_cvc": {
                    "description": "Credit card CVC (for credit_card method)\nexample: 123",
                    "type": "string"
                },
                "card_expiry": {
                    "description": "Credit card expiry date MM/YY (for credit_card method)\nexample: 12/25",
                    "type": "string"
                },
                "card_number": {
                    "description": "Credit card number (for credit_card method)\nexample: 4242424242424242",
                    "type": "string"
                },
                "order_id": {
                    "description": "Order ID - unique identifier for the order\nrequired: true\nexample: order-12345",
                    "type": "string"
                },
                "payment_method": {
                    "description": "Payment method - credit_card, qr_scan, or crypto\nrequired: true\nenum: credit_card,qr_scan,crypto\nexample: credit_card",
                    "type": "string"
                }
            }
        },
        "domain.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Transaction amount\nexample: 250.50",
                    "type": "number"
                },
                "created_at": {
                    "description": "Timestamp when transaction was created\nexample: 2026-01-25T10:30:00Z",
                    "type": "string"
                },
                "currency": {
                    "description": "Currency code\nexample: AED",
                    "type": "string"
                },
                "id": {
                    "description": "Database ID\nexample: 1",
                    "type": "integer"
                },
                "order_id": {
                    "description": "Associated order ID\nexample: order-12345",
                    "type": "string"
                },
                "payment_method": {
                    "description": "Payment method used\nenum: credit_card,qr_scan,crypto\nexample: credit_card",
                    "type": "string"
                },
                "payment_ref": {
                    "description": "External payment gateway reference\nexample: ch_1234567890",
                    "type": "string"
                },
                "status": {
                    "description": "Transaction status\nenum: PENDING,SUCCESS,FAILED\nexample: SUCCESS",
                    "type": "string"
                },
                "transaction_id": {
                    "description": "Unique transaction identifier\nexample: TXN-CC-17000",
                    "type": "string"
                }
            }
        }
    }
}